feat: enable camera combat when setting combat mode

+++ Source/CombatZoneScript.h
@@ -22,5 +22,5 @@ public:
 private:
 	ComponentAudioSource* componentAudio;
 	ComponentRigidBody* componentRigidBody;
-	int enemiesToDefeat;
+	float enemiesToDefeat;
 };
